+++ src/MapApp.tsx
@@ -192,7 +192,7 @@ export default function MapApp() {
     >
       <Sidebar
         active="map"
-        version="v1.0.1"
+        version="v1.0.2"
         hubTargetUuid={hubTargetUuid}
         crossPage
         mobileOpen={drawer.open}
@@ -367,7 +367,7 @@ export default function MapApp() {
           </div>
         </main>
 
-        <Footer version="v1.0.1" config={config} />
+        <Footer version="v1.0.2" config={config} />
       </div>
     </div>
   )